requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect2==0.4.1
Flask==1.1.2
fpdf==1.7.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
Pillow==8.1.0
PyPDF2==1.26.0
reportlab==3.5.59
requests==2.25.1
soupsieve==2.2
svglib==1.0.1
tinycss2==1.1.0
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.1
html5lib==1.1
